Adds shop service to Blazor app
Adds a new service for managing shop data, including fetching, adding, updating, and deleting shop records.
The existing weapon service is updated to use the configuration service to get the base URL for the API, allowing for easier configuration of the API endpoint.
Registers both services as singletons to be available through all the application.